backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
chromadb
pydantic
tenacity
pytest
groq
